Extract out "expandPath" to a separate "path" module
