+++ __tests__/util/fs.js
@@ -1,10 +1,6 @@
 /* @flow */
 
-jest.mock('../../src/util/user-home-dir.js', () => ({
-  default: '/home/foo',
-}));
-
-import {fileDatesEqual, expandPath} from '../../src/util/fs.js';
+import {fileDatesEqual} from '../../src/util/fs.js';
 
 describe('fileDatesEqual', () => {
   const realPlatform = process.platform;
@@ -27,13 +23,6 @@ describe('fileDatesEqual', () => {
       expect(fileDatesEqual(new Date(1491393798834), new Date(1491393798835))).toBeFalsy();
       expect(fileDatesEqual(new Date(1491393700834), new Date(1491393798834))).toBeFalsy();
       expect(fileDatesEqual(new Date(1491393798000), new Date(1491393798835))).toBeFalsy();
-    });
-
-    test('Paths get expanded', () => {
-      expect(expandPath('~/bar/baz/q')).toEqual('/home/foo/bar/baz/q');
-      expect(expandPath('  ~/bar/baz')).toEqual('/home/foo/bar/baz');
-      expect(expandPath('./~/bar/baz')).toEqual('./~/bar/baz');
-      expect(expandPath('~/~/bar/baz')).toEqual('/home/foo/~/bar/baz');
     });
   });
 
@@ -59,13 +48,6 @@ describe('fileDatesEqual', () => {
     test('Milliseconds are ignored when one date has zero milliseconds', () => {
       expect(fileDatesEqual(new Date(1491393798000), new Date(1491393798835))).toBeTruthy();
       expect(fileDatesEqual(new Date(1491393798834), new Date(1491393798000))).toBeTruthy();
-    });
-
-    test('Paths never get expanded', () => {
-      expect(expandPath('~/bar/baz/q')).toEqual('~/bar/baz/q');
-      expect(expandPath('  ~/bar/baz')).toEqual('  ~/bar/baz');
-      expect(expandPath('./~/bar/baz')).toEqual('./~/bar/baz');
-      expect(expandPath('~/~/bar/baz')).toEqual('~/~/bar/baz');
     });
   });
 });
